feat(frontend): :lipstick: create custom button
