--- packages/frontend/src/components/atoms/ButtonCustom.tsx
@@ -1,7 +1,0 @@
-interface Props {
-  content: string;
-}
-
-export const ButtonCustom = ({ content }: Props) => {
-  return <button>{content}</button>;
-};
